move web related code to omod
